Tests TQP and makes some fixes to Stored Procedures and tables

+++ DB/ACVP_DB/acvp/Stored Procedures/VectorSetJsonPut.sql
@@ -1,6 +1,6 @@
 ﻿
 CREATE PROCEDURE [acvp].[VectorSetJsonPut]
-    @VsId BIGINT,
+    @VsId INT,
     @JsonFileType VARCHAR(20),
     @Content VARCHAR(max)
 AS
